move: atom 및 selector 폴더 분리
